Create python rate_things.py
This code will first get the list of all the restaurants in the country specified by the user. It will then get the rating for each restaurant and calculate the average rating. Finally, it will print the average rating to the console.

The code will prompt you to enter the name of the country. Once you have entered the name of the country, the code will print the average rating for restaurants in that country.